Some tests for scales-utils. First tests for the scales; tests for the components were moved to
